Implement source as static class member.
Fixes the need for moving everything to sources.py.  We have
to find out what we want.  It is nice to have build classes
be independent of source urls, and take a source url and
list of patches.  Otoh, it is nice for a build class to
be self-contained.  This is a much more pluggable design
for adding packages than also needing to update a central file.