OpenOffice: build cppuhelper with debug=true.
This should print actual reasons why the install fails.

[It still puzzles me why OpenOffice.org's build and install system
will more often than not just abort without giving the reason of
failure.  Does it only fail for me?  Am I misinformed?  Am I too dense
to interpret the --mostly-- constant error messages?  Are there so
many places of failure that people are too discouraged to fix it?  Is
it slated for a rewrite?  Is making a small fix in ooo-build not worth
it because it won't be upstreamed, and is upstreaming a small fix not
worth the hassle?  Does noone (at upstream) use it?]  confused /me