GIT: only use shallow cloning for non-tracking branches.  Fixes fontconfig.
...and others.  Using GIT to in fact get a historic tar ball, ie,
build from a TAG or COMMITTISH is possibly a bad idea.  You pull all
of git's history, and *need* it, because [with current git] you have
no way of knowing how many revisions back the TAG/COMMITTISH was.

For tracking repositories, we can use shallow clones...which is nice
for a one-time test of GUB, but tracking repositories are usually
the most interesting, and we may want full history there too.

So what we need, is a kind of `git clone --depth-until=COMMITISH',
and switch to tar ball releases as soon an necessary changes from
GIT sources are released (fontconfig, guile, ghostscript, etc).