TODO: musings on priorities.  GUB should be a real distro.
Wouldn't it be nice to have TopGit [or its successor] running and use
that instead of packages.  Upgrading, distributing, installing
[reverting], configuring could all be handled with [Top]Git.  Drop
the .gup tarballs, .rpm, .deb whatnot.

Wouldn't it be nice to have our .spec files use upstream's DSCM if
available?  What if we had a GUB build service running and so provide
a binary GIT archive that provides a runnable GNU/Linux system that is
--within an hour or so-- up to date with every upstream packages that
GUB supports?